generate static js file for i18n
